Move `@SingleTenant` up to annotations package.
