+++ service/src/main/java/org/databiosphere/workspacedataservice/config/TwdsProperties.java
@@ -2,7 +2,7 @@ package org.databiosphere.workspacedataservice.config;
 
 import java.util.Optional;
 import javax.annotation.Nullable;
-import org.databiosphere.workspacedataservice.config.InstanceProperties.SingleTenant;
+import org.databiosphere.workspacedataservice.annotations.SingleTenant;
 import org.databiosphere.workspacedataservice.shared.model.WorkspaceId;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
